ublk_drv: define macros for recovery feature and check them
Define some macros for recovery feature. Especially define a new state:
UBLK_S_DEV_RECOVERING which implies the ublk_device is recovering.

For now, we do not allow STOP_DEV while we are in UBLK_S_DEV_RECOVERING.
This means that we must assign a new ubq_daemon for each ubq.

Also modify checks on state in START_DEV and SET_PARAMS because now we
have three states.

Signed-off-by: ZiyangZhang <ZiyangZhang@linux.alibaba.com>