ublk_drv: do not run monitor_work while ub's state is QUIESCED
We cannot run monitor_work in this state because START_USER_RECOVERY
may reset ubq_daemon to NULL and wait for new process. Besides,
monitor_work is not necessary in QUIESCED state since we have already
scheduled quiesce_work and quiesce all ubqs.

monitor_work should not run if state it either DEAD or QUIESCED. And we
cancel it in START_USER_RECOVERY and re-schedule it in
END_USER_RECOVERY.

Signed-off-by: ZiyangZhang <ZiyangZhang@linux.alibaba.com>